topbar breakdown into widgets
