+++ components/dashboard/topbarWidget/styles.ts
@@ -1,0 +1,20 @@
+// components/layout/topbar/topbarWidgets/styles.ts
+"use client";
+
+// Define base colors for reuse, makes theme changes easier
+export const textColor = "text-indigo-100 dark:text-gray-200";
+export const hoverTextColor = "hover:text-white dark:hover:text-white";
+export const iconColor = "text-indigo-200 dark:text-gray-400";
+export const bgColor = "bg-indigo-700 dark:bg-gray-950";
+export const borderColor = "border-indigo-900 dark:border-gray-800";
+export const hoverBgColor = "hover:bg-indigo-600 dark:hover:bg-gray-800";
+export const separatorColor = "bg-indigo-500 dark:bg-gray-700";
+
+// Dropdown specific colors
+export const dropdownBgColor = "bg-white dark:bg-gray-900";
+export const dropdownBorderColor = "border-gray-200 dark:border-gray-700";
+export const dropdownSeparatorColor = "bg-gray-200 dark:bg-gray-700";
+export const dropdownFocusBgColor = "focus:bg-gray-100 dark:focus:bg-gray-800";
+export const dropdownTextColor = "text-gray-700 dark:text-gray-300";
+export const dropdownLabelColor = "text-gray-900 dark:text-gray-100";
+export const dropdownMutedColor = "text-gray-500 dark:text-gray-400";
